Improve bigfiles situation on mysql branch (#42)
* Extract out test case to its own class

* Extract MoverTestCase from TableValidator

* Move more content over to MoverTestCase

* Start extracting things out of main logging typestub

* Extract out RecordsTableTimezoneValidator

* Break out logging typestubs further

* Allow bigfiles slip

350: itest
317: records_mover/records/schema/field/__init__.py
300: setup.py

Reduce total number of bigfiles violations to 897 or below!

Tasks: TOP => bigfiles
(See full trace by running task with --trace)
Existing violations: 897
Found 967 bigfiles violations

* Fix long lines

* Ratchet

* Ratchet
