+++ types/stubs/logging/__init__.pyi
@@ -1,4 +1,14 @@
+__all__ = [
+    'Logger',
+    'Filterer',
+    'LogRecord',
+]
+
 from typing import Any, Optional
+from .log_record import LogRecord
+from .logger import Logger
+from .filterer import Filterer
+
 
 raiseExceptions: bool
 CRITICAL: int
@@ -15,45 +25,6 @@ def getLevelName(level: Any): ...
 
 
 def addLevelName(level: Any, levelName: Any) -> None: ...
-
-
-class LogRecord:
-    name: Any = ...
-    msg: Any = ...
-    args: Any = ...
-    levelname: Any = ...
-    levelno: Any = ...
-    pathname: Any = ...
-    filename: Any = ...
-    module: Any = ...
-    exc_info: Any = ...
-    exc_text: Any = ...
-    stack_info: Any = ...
-    lineno: Any = ...
-    funcName: Any = ...
-    created: Any = ...
-    msecs: Any = ...
-    relativeCreated: Any = ...
-    thread: Any = ...
-    threadName: Any = ...
-    processName: Any = ...
-    process: Any = ...
-
-    def __init__(
-        self,
-        name: Any,
-        level: Any,
-        pathname: Any,
-        lineno: Any,
-        msg: Any,
-        args: Any,
-        exc_info: Any,
-        func: Optional[Any] = ...,
-        sinfo: Optional[Any] = ...,
-        **kwargs: Any
-    ) -> None: ...
-
-    def getMessage(self): ...
 
 
 def setLogRecordFactory(factory: Any) -> None: ...
@@ -131,14 +102,6 @@ class Filter:
     name: Any = ...
     nlen: Any = ...
     def __init__(self, name: str = ...) -> None: ...
-    def filter(self, record: Any): ...
-
-
-class Filterer:
-    filters: Any = ...
-    def __init__(self) -> None: ...
-    def addFilter(self, filter: Any) -> None: ...
-    def removeFilter(self, filter: Any) -> None: ...
     def filter(self, record: Any): ...
 
 
@@ -222,53 +185,6 @@ class Manager:
     def getLogger(self, name: str) -> "Logger": ...
     def setLoggerClass(self, klass: Any) -> None: ...
     def setLogRecordFactory(self, factory: Any) -> None: ...
-
-
-class Logger(Filterer):
-    name: Any = ...
-    level: Any = ...
-    parent: Any = ...
-    propagate: bool = ...
-    handlers: Any = ...
-    disabled: bool = ...
-    def __init__(self, name: Any, level: Any = ...) -> None: ...
-    def setLevel(self, level: Any) -> None: ...
-    def debug(self, msg: str, *args: Any, **kwargs: Any) -> None: ...
-    def info(self, msg: str, *args: Any, **kwargs: Any) -> None: ...
-    def warning(self, msg: str, *args: Any, **kwargs: Any) -> None: ...
-    def warn(self, msg: str, *args: Any, **kwargs: Any) -> None: ...
-    def error(self, msg: str, *args: Any, **kwargs: Any) -> None: ...
-
-    def exception(
-        self, msg: str, *args: Any, exc_info: bool = ..., **kwargs: Any
-    ) -> None: ...
-    def critical(self, msg: str, *args: Any, **kwargs: Any) -> None: ...
-    fatal: Any = ...
-    def log(self, level: Any, msg: str, *args: Any, **kwargs: Any) -> None: ...
-    def findCaller(self, stack_info: bool = ..., stacklevel: int = ...): ...
-
-    def makeRecord(
-        self,
-        name: Any,
-        level: Any,
-        fn: Any,
-        lno: Any,
-        msg: Any,
-        args: Any,
-        exc_info: Any,
-        func: Optional[Any] = ...,
-        extra: Optional[Any] = ...,
-        sinfo: Optional[Any] = ...,
-    ): ...
-    def handle(self, record: Any) -> None: ...
-    def addHandler(self, hdlr: Any) -> None: ...
-    def removeHandler(self, hdlr: Any) -> None: ...
-    def hasHandlers(self): ...
-    def callHandlers(self, record: Any) -> None: ...
-    def getEffectiveLevel(self): ...
-    def isEnabledFor(self, level: Any): ...
-    def getChild(self, suffix: Any): ...
-    def __reduce__(self): ...
 
 
 class RootLogger(Logger):
