+++ types/stubs/logging/filterer.py
@@ -1,0 +1,9 @@
+from typing import Any
+
+
+class Filterer:
+    filters: Any = ...
+    def __init__(self) -> None: ...
+    def addFilter(self, filter: Any) -> None: ...
+    def removeFilter(self, filter: Any) -> None: ...
+    def filter(self, record: Any): ...
