+++ types/stubs/logging/log_record.pyi
@@ -1,0 +1,40 @@
+from typing import Any, Optional
+
+
+class LogRecord:
+    name: Any = ...
+    msg: Any = ...
+    args: Any = ...
+    levelname: Any = ...
+    levelno: Any = ...
+    pathname: Any = ...
+    filename: Any = ...
+    module: Any = ...
+    exc_info: Any = ...
+    exc_text: Any = ...
+    stack_info: Any = ...
+    lineno: Any = ...
+    funcName: Any = ...
+    created: Any = ...
+    msecs: Any = ...
+    relativeCreated: Any = ...
+    thread: Any = ...
+    threadName: Any = ...
+    processName: Any = ...
+    process: Any = ...
+
+    def __init__(
+        self,
+        name: Any,
+        level: Any,
+        pathname: Any,
+        lineno: Any,
+        msg: Any,
+        args: Any,
+        exc_info: Any,
+        func: Optional[Any] = ...,
+        sinfo: Optional[Any] = ...,
+        **kwargs: Any
+    ) -> None: ...
+
+    def getMessage(self): ...
